requirements.txt
python-engineio==3.12.1
python-socketio==4.5.1
PyYAML==5.3.1
requests==2.22.0
urllib3==1.25.7
